requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
elevenlabs==0.2.26
python-multipart==0.0.6
pydantic==2.5.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
openai==1.3.0
aiofiles==0.23.2
loguru==0.7.2

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Payment processing
stripe==7.8.0

# Additional utilities
email-validator==2.1.0
requests==2.31.0

# Testing dependencies
pytest==7.4.3
httpx==0.25.2

# NOTE: Do not put secrets or API keys in this file. Use a .env file for sensitive information.
